Use float64 in the ZFS collector
ZFS metrics can also be unsigned 64-bit integers that won't fit in
int64 and causes the whole collector to fail.

+++ collector/zfs.go
@@ -75,7 +75,7 @@ func (s zfsSysctl) metricName() string {
 	return strings.Replace(parts[len(parts)-1], "-", "_", -1)
 }
 
-func (c *zfsCollector) constSysctlMetric(subsystem string, sysctl zfsSysctl, value int64) prometheus.Metric {
+func (c *zfsCollector) constSysctlMetric(subsystem string, sysctl zfsSysctl, value float64) prometheus.Metric {
 	metricName := sysctl.metricName()
 
 	return prometheus.MustNewConstMetric(
@@ -86,11 +86,11 @@ func (c *zfsCollector) constSysctlMetric(subsystem string, sysctl zfsSysctl, val
 			nil,
 		),
 		prometheus.UntypedValue,
-		float64(value),
+		value,
 	)
 }
 
-func (c *zfsCollector) constPoolMetric(poolName string, sysctl zfsSysctl, value int64) prometheus.Metric {
+func (c *zfsCollector) constPoolMetric(poolName string, sysctl zfsSysctl, value float64) prometheus.Metric {
 	metricName := sysctl.metricName()
 
 	return prometheus.MustNewConstMetric(
@@ -101,7 +101,7 @@ func (c *zfsCollector) constPoolMetric(poolName string, sysctl zfsSysctl, value 
 			nil,
 		),
 		prometheus.UntypedValue,
-		float64(value),
+		value,
 		poolName,
 	)
 }
